agents/agent-interactive/requirements.txt
# --- Core runtime ---
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
openai==0.27.4

# --- Optional / later phases ---
faiss-cpu==1.8.0      # or chromadb==0.5.0  (choose one later)
kafka-python==2.0.2
# ----- kafka testing -----
python-dotenv==1.0.0
# -- embedding locally installed --
# Force pip to pick CPU-only wheels from PyTorch index
--index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://pypi.org/simple
torch==2.4.1+cpu          # must come before sentence-transformers to bypass CUDA/NVIDIA dependencies
sentence-transformers==3.0.1
